+++ eras/shelley/impl/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Version history for `cardano-ledger-shelley`
 
+## 1.6.1.0
+
+* Add `FromJSON` instance for `Constitution`
+
 ## 1.6.0.0
 
 * Add lenses for `UTxOEnv` #3688
-* Add `FromJSON` instance for `Constitution`
 * Add `getTotalTxDepositsBody` to `ShelleyEraTxBody`
 * Add `obligationGovState` to `EraGov`
 * Add `ToExpr` instance to `Constitution`